Make bottom and top shadow into two separate views
